web_monitor_rag/requirements.txt
Flask
requests
beautifulsoup4
psycopg2-binary
chromadb
numpy
